remove old InfoBoxMeasurement in commons/cismap lib, clean up types in measurements
