+++ libraries/commons/measurements/src/lib/components/MeasurementTitle.tsx
@@ -1,5 +1,15 @@
 import { useState, useEffect } from "react";
-import { MeasurementTitleProps } from "./types.d";
+
+type MeasurementTitleProps = {
+  title: string;
+  shapeId: number | string;
+  order: number;
+  updateTitleMeasurementById: (shapeId: number | string, title: string) => void;
+  setUpdateMeasurementStatus: (status: boolean) => void;
+  isCollapsed?: boolean;
+  collapsedContent?: string;
+  editable?: boolean;
+};
 
 const MeasurementTitle = ({
   title,
