+++ libraries/commons/measurements/src/lib/types/SnappingPoint.d.ts
@@ -1,0 +1,10 @@
+export type SnappingPoint = {
+  coordinates: [number, number]; // [lng, lat]
+  sourceId: string;
+  distance?: number; // Calculated later
+  metadata?: {
+    featureId?: string;
+    shapeId?: string | number;
+    geometryType?: string;
+  };
+};
